requirements.txt
# Certificate Generation System - Core Dependencies
flask==2.3.3
pillow==10.0.1
qrcode[pil]==7.4.2
reportlab==4.0.4
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
gunicorn==21.2.0
python-dotenv==1.0.0